Vendor request to query frame draw timing.
Since the interrupt endpoint behaves as FIFO, it has to be drained
at least once to get recent frame draw timing information. With a
control transfer the information can be latched when the setup
request is received, ensuring recent information is always available.

This also doesn't require to wait for the next transfer opportunity
of the interrupt endpoint, which is annoying when you only have blocking
call available in Python.