Use an EEPROM stored group identifier.
Instead of using a (flawed) heuristic algorithm to group display
controllers, rely on the controllers themselves to advertise how they
should be grouped.
As this may result in gaps in the actually supported information range,
some extra buffer is provided. This extra space is can be written to,
but is unused.