Support for reversing strip segment orientation.
The Münster group has placed the electronics in the upper part
of the display, effectively reversing the order of all LED strip
segments. This change adds a new flag to the EEPROM that indicates
if the first segment is reversed.
For uninitialised EEPROM, this means it defaults to the behaviour
like in the original display design, with the electronics in the
bottom part of the display.